Move methods from Story provider to controller
