+++ lib/Alchemy/Phrasea/Application.php
@@ -12,7 +12,6 @@
 namespace Alchemy\Phrasea;
 
 use Alchemy\Geonames\GeonamesServiceProvider;
-use Alchemy\Phrasea\ControllerProvider\Prod\Story;
 use Alchemy\Phrasea\ControllerProvider\Prod\Tools;
 use Alchemy\Phrasea\ControllerProvider\Prod\Tooltip;
 use Alchemy\Phrasea\ControllerProvider\Prod\TOU;
@@ -315,6 +314,7 @@ class Application extends SilexApplication
             'Alchemy\Phrasea\ControllerProvider\Prod\Record' => [],
             'Alchemy\Phrasea\ControllerProvider\Prod\Root' => [],
             'Alchemy\Phrasea\ControllerProvider\Prod\Share' => [],
+            'Alchemy\Phrasea\ControllerProvider\Prod\Story' => [],
             'Alchemy\Phrasea\ControllerProvider\Datafiles' => [],
             'Alchemy\Phrasea\ControllerProvider\Lightbox' => [],
             'Alchemy\Phrasea\ControllerProvider\MediaAccessor' => [],
@@ -625,7 +625,6 @@ class Application extends SilexApplication
         $this->mount('/login/', new Login());
         $this->mount('/developers/', new Developers());
 
-        $this->mount('/prod/story', new Story());
         $this->mount('/prod/WorkZone', new WorkZone());
         $this->mount('/prod/lists', new UsrLists());
         $this->mount('/prod/TOU/', new TOU());
@@ -681,6 +680,7 @@ class Application extends SilexApplication
             '/prod/records/movecollection' => 'Alchemy\Phrasea\ControllerProvider\Prod\MoveCollection',
             '/prod/records/property'       => 'Alchemy\Phrasea\ControllerProvider\Prod\Property',
             '/prod/share/'                 => 'Alchemy\Phrasea\ControllerProvider\Prod\Share',
+            '/prod/story'                  => 'Alchemy\Phrasea\ControllerProvider\Prod\Story',
             '/prod/'                       => 'Alchemy\Phrasea\ControllerProvider\Prod\Root',
             '/setup'                       => 'Alchemy\Phrasea\ControllerProvider\Setup',
         ];
